pyproject.toml
[project]
name = "epg-irib"
version = "0.1.0"
description = "Generate XMLTV EPG data for IRIB channels from Sepehr TV API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests-oauthlib>=1.3.0",
]

[project.scripts]
epg-irib = "generate_epg:main"
